+++ fwdpp/ts/table_collection.hpp
@@ -281,7 +281,7 @@ namespace fwdpp
             {
                 for (auto i : node_ids)
                     {
-                        if (i >= node_table.size())
+                        if (static_cast<std::size_t>(i) >= node_table.size())
                             {
                                 throw std::invalid_argument(
                                     "node id larger than node table size");
